Apply changes from code review and improve benchmark
